+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Version 0.2
 
 * Error messages are translated into error events, and do not return an error
   code.
+* The client and server objects accept a direction parameter for
+  close_connection, obsoleting then eed to provide an empty buffer to read_data.
 
 Version 0.1
 -----------